Move `CurrencyFormatter` to currency_formatter.rs
